Move methods to individual package
