+++ perun/check/meson.build
@@ -2,20 +2,14 @@ perun_check_dir = perun_dir / 'check'
 
 perun_check_files = files(
     '__init__.py',
-    'average_amount_threshold.py',
-    'best_model_order_equality.py',
-    'exclusive_time_outliers.py',
     'factory.py',
-    'fast_check.py',
     'general_detection.py',
-    'integral_comparison.py',
-    'linear_regression.py',
-    'local_statistics.py',
     'nonparam_helpers.py',
-    'polynomial_regression.py',
 )
 
 py3.install_sources(
     perun_check_files,
     subdir: perun_check_dir,
 )
+
+subdir('methods')
