+++ perun/check/methods/average_amount_threshold.py
