+++ perun/check/methods/integral_comparison.py
