+++ perun/check/methods/local_statistics.py
