+++ perun/check/methods/polynomial_regression.py
